backend/CAcceptJob/requirements.txt
Flask==2.1.1
Flask-Cors==5.0.1
requests==2.26.0
pika==1.2.0
kafka-python==2.0.2
Werkzeug==2.1.2